+++ src/types/ConnectionProfile.ts
@@ -1,4 +1,4 @@
-import { NetworkSettings, OrgConfig, PeerConfig } from "./FabloConfigExtended";
+import {ChannelConfig, NetworkSettings, OrgConfig, PeerConfig} from "./FabloConfigExtended";
 
 interface ConnectionProfile {
   name: string;
@@ -41,15 +41,58 @@ interface HttpOptions {
   verify: boolean;
 }
 
+interface HyperledgerExplorerConnectionProfile {
+  name: string;
+  description: string;
+  version: string;
+  peers: { [key: string]: Peer };
+  client: HyperledgerExplorerClient;
+  organizations: { [key: string]: HyperledgerExplorerOrganization };
+  channels: { [name: string]: Channel };
+}
+
+interface HyperledgerExplorerClient extends Client {
+  tlsEnable: boolean;
+  adminCredential: {
+    id: string;
+    password: string;
+  };
+  enableAuthentication: boolean;
+  connection: {
+    timeout: {
+      peer: {
+        endorser: number;
+      };
+      orderer: number;
+    };
+  };
+}
+
+interface HyperledgerExplorerOrganization {
+  mspid: string;
+  peers: Array<string>;
+  adminPrivateKey: {
+    path: string;
+  };
+  signedCert: {
+    path: string;
+  };
+}
+
+interface Channel {
+  peers: { [address: string]: unknown };
+}
+
 function createPeers(isTls: boolean, rootPath: string, orgs: OrgConfig[]): { [key: string]: Peer } {
   const peers: { [key: string]: Peer } = {};
   orgs.forEach((o: OrgConfig) => {
     o.anchorPeers.forEach((p: PeerConfig) => {
       if (isTls) {
         peers[p.address] = {
-          url: `grpcs://localhost:${p.port}`,
+          url: p.fullAddress,
+          // url: `grpcs://localhost:${p.port}`,
           tlsCACerts: {
-            path: `${rootPath}/fablo-target/fabric-config/crypto-config/peerOrganizations/${o.domain}/peers/${p.address}/tls/ca.crt`,
+            path: `${rootPath}/peerOrganizations/${o.domain}/peers/${p.address}/tls/ca.crt`, // todo is it good?
           },
           grpcOptions: {
             "ssl-target-name-override": p.address,
@@ -57,7 +100,8 @@ function createPeers(isTls: boolean, rootPath: string, orgs: OrgConfig[]): { [ke
         };
       } else {
         peers[p.address] = {
-          url: `grpc://localhost:${p.port}`,
+          url: p.fullAddress,
+          // url: `grpc://localhost:${p.port}`,
         };
       }
     });
@@ -73,10 +117,11 @@ function certificateAuthorities(
   if (isTls) {
     return {
       [org.ca.address]: {
-        url: `http://localhost:${org.ca.exposePort}`,
+        url: org.ca.fullAddress,
+        // url: `http://localhost:${org.ca.exposePort}`,
         caName: org.ca.address,
         tlsCACerts: {
-          path: `${rootPath}/fablo-target/fabric-config/crypto-config/peerOrganizations/${org.domain}/ca/${org.ca.address}-cert.pem`,
+          path: `${rootPath}/peerOrganizations/${org.domain}/ca/${org.ca.address}-cert.pem`,
         },
         httpOptions: {
           verify: false,
@@ -86,7 +131,8 @@ function certificateAuthorities(
   }
   return {
     [org.ca.address]: {
-      url: `http://localhost:${org.ca.exposePort}`,
+      url: org.ca.fullAddress,
+      // url: `http://localhost:${org.ca.exposePort}`,
       caName: org.ca.address,
       httpOptions: {
         verify: false,
@@ -95,12 +141,23 @@ function certificateAuthorities(
   };
 }
 
+function createChannels(org: OrgConfig, channels: ChannelConfig[]): { [name: string]: Channel } {
+  const peers: { [address: string]: unknown } = {};
+  const cs: { [name: string]: Channel } = {};
+  channels.forEach((channel) => {
+    if (channel.orgs.map((o) => o.name).indexOf(org.name) != -1) {
+      cs["my-channel1"] = { peers: peers };
+    }
+  });
+  return cs;
+}
+
 export function createConnectionProfile(
   networkSettings: NetworkSettings,
   org: OrgConfig,
   orgs: OrgConfig[],
 ): ConnectionProfile {
-  const rootPath = networkSettings.paths.chaincodesBaseDir;
+  const rootPath = `${networkSettings.paths.chaincodesBaseDir}/fablo-target/fabric-config/crypto-config`;
   const peers = createPeers(networkSettings.tls, rootPath, orgs);
   return {
     name: `fablo-test-network-${org.name.toLowerCase()}`,
@@ -118,5 +175,51 @@ export function createConnectionProfile(
     },
     peers: peers,
     certificateAuthorities: certificateAuthorities(networkSettings.tls, rootPath, org),
+  };
+}
+
+export function createHyperledgerExplorerConnectionProfile(
+  networkSettings: NetworkSettings,
+  org: OrgConfig,
+  orgs: OrgConfig[],
+  channels: ChannelConfig[],
+): HyperledgerExplorerConnectionProfile {
+  const rootPath = "/tmp/crypto";
+  const peers = createPeers(networkSettings.tls, rootPath, orgs);
+  return {
+    name: `fablo-test-network-${org.name.toLowerCase()}`,
+    description: `Connection profile for Hyperledger Explorer in Fablo network`,
+    version: "1.0.0",
+    client: {
+      organization: org.name,
+      tlsEnable: networkSettings.tls,
+      enableAuthentication: true,
+      adminCredential: {
+        id: "admin",
+        password: "adminpw",
+      },
+      connection: {
+        timeout: {
+          peer: {
+            endorser: 300,
+          },
+          orderer: 300,
+        },
+      },
+    },
+    organizations: {
+      [org.name]: {
+        mspid: org.mspName,
+        peers: Object.keys(peers),
+        adminPrivateKey: {
+          path: `${rootPath}/peerOrganizations/${org.domain}/users/Admin@${org.domain}/msp/keystore/priv-key.pem`,
+        },
+        signedCert: {
+          path: `${rootPath}/peerOrganizations/${org.domain}/users/Admin@${org.domain}/msp/signcerts/Admin@${org.domain}-cert.pem`,
+        },
+      },
+    },
+    peers: peers,
+    channels: createChannels(org, channels),
   };
 }